safari-600.7-branch/LayoutTests/media/video-controller-currentTime-expected.txt


Test that seeking video with media controller throws invalid state exception.

EVENT(canplaythrough)
TEST(video.currentTime = 3) THROWS(DOMException.INVALID_STATE_ERR) OK
END OF TEST
